Add: los jugadores pueden agarrar armas

+++ server/game_logic/player/position/player_position.h
@@ -33,7 +33,7 @@ public:
     void occupy(Coordinate&);
     void free_occupied();
     std::vector<Coordinate> get_occupied();
-    uint8_t get_facing_direction();
+    int get_facing_direction();
     void released_jump();
     void released_w();
     bool is_aiming_up();
